+++ packages/components/molecules/f-media-element/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+
+v3.0.0
+-----------------------------
+*June 20, 2022*
+
+### Changed
+- Update to `@use` and `@forward` SASS syntax
+
+
 v2.0.1
 ------------------------------
 *Jun 9, 2022*