+++ packages/components/pages/f-registration/CHANGELOG.md
@@ -4,12 +4,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-v2.3.0
+v2.4.0
 ------------------------------
-*June 17, 2022*
+*June 21, 2022*
 
 ### Changed
 - Update `f-link` to v3.1.0.
+
+
+v2.3.0
+------------------------------
+*June 20, 2022*
+
+### Changed
+- Bumped `f-form-field` version to add `required` fields.
 
 
 v2.2.0